Update README.md
This project, Sound Store, is an e-commerce platform focused on selling headphones and sound-related products. It includes features like a chatbot for customer support, live technical assistance, and a user-friendly interface for browsing and managing products. The platform is built using PHP, MySQL, and Bootstrap for a seamless shopping experience.